Fixed: All but shield upgrades work and UI is fixed as well

+++ 2D_TurnBased_Template/Assets/Code/Base Classes/BaseCharacter.cs
@@ -10,7 +10,7 @@ public class BaseCharacter : MonoBehaviour// need to move melee and rage values 
     public float CharacterMaxHealthLevel;
 
     [Header("Range Dmg")]
-    public int RangeDamg;
+    public float RangeDamg;
     [Header("Souls/XP")]
     public int Souls;
     [Header("Booleans")]
